+++ budget/model/valueObject/budget.go
@@ -1,7 +1,7 @@
 package valueobject
 
 type Budget struct {
-	Spent      float32
-	Total     float32
-	Remaining float32
+	Spent     float64
+	Total     float64
+	Remaining float64
 }
